# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "API Documentation for my Laravel application",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get the list of companies for the authenticated user",
                "description": "Returns a list of companies that belong to the currently authenticated user",
                "operationId": "getUserCompanies",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Company"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Create a new company",
                "description": "Creates a new company associated with the authenticated user. Restores a soft-deleted company if the same INN exists.",
                "operationId": "createCompany",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "inn",
                                    "title"
                                ],
                                "properties": {
                                    "inn": {
                                        "description": "INN of the company",
                                        "type": "string",
                                        "example": "123456789012"
                                    },
                                    "title": {
                                        "description": "Title of the company",
                                        "type": "string",
                                        "example": "My Company"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Company with this INN already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Company with this INN already exists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/companies/{company}": {
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update a company",
                "description": "Updates the title of a company. Only the owner of the company (authenticated user) can update it.",
                "operationId": "updateCompany",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "ID of the company to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Updated title of the company",
                                        "type": "string",
                                        "example": "Updated Company Title"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Delete a company",
                "description": "Soft deletes a company. Only the owner of the company (authenticated user) can delete it.",
                "operationId": "deleteCompany",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "ID of the company to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Company deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Company": {
                "required": [
                    "id",
                    "inn",
                    "title",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "description": "Company ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "inn": {
                        "description": "INN of the company",
                        "type": "string"
                    },
                    "title": {
                        "description": "Title of the company",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "ID of the user who owns the company",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Companies",
            "description": "API Endpoints for Managing Companies"
        }
    ]
}